Fix test failure in DDBasicLoggingTests
There was a comment, in line 55 indicating that the block would be called twice
for every log. Appearantly this isn't the case anymore.

Note: I do not fully understand what is going on there and why the behavior
seems to have changed.
I only observed the current behavior to be that the block is called exactly
once for every log.

@bpoplauschi please double check this commit